Move NewQualityProfile to BETA and split
