+++ MG.Sonarr-DotNet/Cmdlets/Connection/Exclude/ConnectInstance-BETA.cs
@@ -1,0 +1,20 @@
+﻿using MG.Sonarr.Results;
+using System;
+using System.Management.Automation;
+
+namespace MG.Sonarr.Cmdlets
+{
+    public partial class ConnectInstance
+    {
+
+        // BETA - for New-QualityDefinition
+        protected override void EndProcessing()
+        {
+            string qStr = base.TryGetSonarrResult("/qualitydefinition");
+            if (!string.IsNullOrEmpty(qStr))
+            {
+                Context.Qualities = SonarrHttp.ConvertToSonarrResults<QualityDefinition>(qStr, out bool iso);
+            }
+        }
+    }
+}
