+++ MG.Sonarr-DotNet/Cmdlets/Profiles/Exclude/NewQualityProfile.cs
