+++ MG.Sonarr-DotNet/MG.Sonarr-DotNet.csproj
@@ -74,12 +74,14 @@
     <Compile Include="Cmdlets\Commands\InvokeSeriesSearch.cs" />
     <None Include="Cmdlets\Connection\Exclude\CertValidity-Core.cs" />
     <Compile Include="Cmdlets\Connection\Exclude\CertValidity-DotNet.cs" />
-    <Compile Include="Cmdlets\Profiles\NewQualityProfile.cs" />
+    <None Include="Cmdlets\Profiles\Exclude\NewQualityProfile.cs" />
+    <None Include="Cmdlets\Connection\Exclude\ConnectInstance-BETA.cs" />
     <Compile Include="Cmdlets\Update\GetUpdate.cs" />
+    <None Include="Components\BETA\SonarrContext-BETA.cs" />
     <Compile Include="Components\SonarrHttp.cs" />
     <Compile Include="Objects\Interfaces\ISonarrUrl.cs" />
     <Compile Include="Objects\SonarrUrl.cs" />
-    <Compile Include="Results\Quality\QualityDefinition.cs" />
+    <None Include="Results\Quality\QualityDefinition.cs" />
     <Compile Include="Results\Updates\Update.cs" />
     <Compile Include="Cmdlets\Connection\ConnectInstance.cs" />
     <Compile Include="Cmdlets\Episodes\GetEpisode.cs" />
